requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
openai>=1.0.0
python-dotenv
pyarrow